requirements/development.txt
-r base.txt
django-debug-toolbar>=4.2.0
django-extensions>=3.2.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
